refactor : Repository용 DTO 생성 후 매핑
