+++ src/main/java/com/opus/opus/modules/team/domain/dao/TeamRepository.java
@@ -1,6 +1,5 @@
 package com.opus.opus.modules.team.domain.dao;
 
-import com.opus.opus.modules.contest.application.dto.response.ContestRankingResponse;
 import com.opus.opus.modules.team.domain.Team;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -15,7 +14,8 @@ public interface TeamRepository extends JpaRepository<Team, Long> {
     boolean existsByTrackId(final Long trackId);
 
     List<Team> findAllByContestId(final Long contestId);
-    @Query("SELECT new com.opus.opus.modules.contest.application.dto.response.ContestRankingResponse(" +
+
+    @Query("SELECT new com.opus.opus.modules.team.domain.dao.TeamRankingResult(" +
             "team.id, team.teamName, team.projectName, track.trackName, COUNT(vote.id)) " +
             "FROM Team team " +
             "LEFT JOIN TeamVote vote ON vote.team = team AND vote.isVoted = true " +
@@ -23,5 +23,5 @@ public interface TeamRepository extends JpaRepository<Team, Long> {
             "WHERE team.contestId = :contestId " +
             "GROUP BY team.id, team.teamName, team.projectName, track.trackName " +
             "ORDER BY COUNT(vote.id) DESC")
-    List<ContestRankingResponse> findTeamRankingByContestId(Long contestId); // 특정 대회에 속한 모든 팀을, 투표 수 기준 내림차순으로 조회 (투표 수 0인 팀도 포함)
+    List<TeamRankingResult> findTeamRankingByContestId(Long contestId); // 특정 대회에 속한 모든 팀을, 투표 수 기준 내림차순으로 조회 (투표 수 0인 팀도 포함)
 }
